Permissions for system commands.

+++ commands/owner/sync.js
@@ -6,34 +6,35 @@ module.exports = {
     description: "sync data to database",
     commandOptions: null,
     async execute(interaction) {  
-        if (interaction.member.user.id !== "382918201241108481") {
-            let embed = new Discord.MessageEmbed()
-                .setTitle("Permission denied.")
-                .setDescription(`Only my developer can use this command...`);
-            client.api.interactions(interaction.id, interaction.token).callback.post({
-                data: {
-                    type: 4,
-                    data: {
-                        embeds: [embed],
-                        ephemeral: true
-                    }
-                }
-            });
-            return;
-        }
-        
-        client.guilds.cache.forEach(guild => {
-            guild.members.cache.forEach(member => {
-                if (!member.user.bot) {
-                    client.Database.deleteMember(guild.id, member.id);
-                    client.Database.fetchMember(member.id, guild.id).then((result) => {
-                        if (result.iconURL === undefined)
-                            client.logger.loger('WARN', `${result.id}/${result.username} invalid iconURL constant.`);
-                        else
-                            client.logger.loger('INFO', `${result.id}/${result.username} successfully added in the database.`);
+        await client.Database.isDeveloper(interaction.member.user.id, result => {
+            if (result.isDev) {
+                client.guilds.cache.forEach(guild => {
+                    guild.members.cache.forEach(member => {
+                        if (!member.user.bot) {
+                            client.Database.deleteMember(guild.id, member.id);
+                            client.Database.fetchMember(member.id, guild.id).then((result) => {
+                                if (result.iconURL === undefined)
+                                    client.logger.loger('WARN', `${result.id}/${result.username} invalid iconURL constant.`);
+                                else
+                                    client.logger.loger('INFO', `${result.id}/${result.username} successfully added in the database.`);
+                            });
+                        }
                     });
-                }
-            });
-        });
+                });
+            } else {
+                let embed = new Discord.MessageEmbed()
+                    .setTitle("Permission denied.")
+                    .setDescription(`Only my developer can use this command...`);
+                client.api.interactions(interaction.id, interaction.token).callback.post({
+                    data: {
+                        type: 4,
+                        data: {
+                            embeds: [embed],
+                            ephemeral: true
+                        }
+                    }
+                });
+            }
+        }); 
     }
 }
